+++ src/test/php/net/xp_framework/unittest/annotations/BrokenAnnotationTest.class.php
@@ -8,6 +8,7 @@ use lang\reflect\ClassParser;
  *
  * @see   rfc://0185
  * @see   https://github.com/xp-framework/xp-framework/pull/328
+ * @see   https://github.com/xp-framework/xp-framework/issues/313
  */
 class BrokenAnnotationTest extends \unittest\TestCase {
 
@@ -149,5 +150,25 @@ class BrokenAnnotationTest extends \unittest\TestCase {
   #[@test, @expect(class= 'lang.ClassFormatException', withMessage= '/In `.+`: Parse error/')]
   public function function_without_closing_curly() {
     $this->parse('#[@value(function() {)]');
+  }
+
+  #[@test, @expect(class= 'lang.ClassFormatException', withMessage= '/Parse error: Unexpected ","/')]
+  public function multi_value() {
+    $this->parse("#[@xmlmapping('hw_server', 'server')]");
+  }
+
+  #[@test, @expect(class= 'lang.ClassFormatException', withMessage= '/Parse error: Unexpected ","/')]
+  public function multi_value_without_whitespace() {
+    $this->parse("#[@xmlmapping('hw_server','server')]");
+  }
+
+  #[@test, @expect(class= 'lang.ClassFormatException', withMessage= '/Parse error: Unexpected ","/')]
+  public function multi_value_with_variable_types_backwards_compatibility() {
+    $this->parse("#[@xmlmapping('hw_server', TRUE)]");
+  }
+
+  #[@test, @expect(class= 'lang.ClassFormatException', withMessage= '/Parse error: Unexpected ","/')]
+  public function parsingContinuesAfterMultiValue() {
+    $this->parse("#[@xmlmapping('hw_server', 'server'), @restricted]");
   }
 }